dumped hardware-interfacing code into a separate package from the rest of the program code
