+++ pinger_finder/oneclk.py
@@ -1,6 +1,6 @@
 from sys import argv
 
-from ADC import ADS7865
+from bbbio.ADC import ADS7865
 import QikSampNPlt
 import locPinger
