+++ pinger_finder/trackPinger.py
@@ -1,7 +1,7 @@
 import logging
 
 import getHeading
-import boot
+from bbbio import boot
 
 logging.basicConfig(level=logging.info, format='%(asctime)s - %(levelname)s - %(message)s')
